Extract favorites hydrator from nostr-auth.tsx
Moves favoriteFromPodcast / resolveGuidToFavorite / hydrateFavorites
into lib/nostr/favorites-hydrator.ts and re-exports via the nostr
barrel. NostrAuth shrinks 324 -> 216 lines and the PI-circuit-breaker
probe-first pattern is now reusable.
